tech-signal-sales-brief/pyproject.toml
[project]
name = "tech-signal-sales-brief"
version = "0.1.0"
description = "A multi-agent system for monitoring and content creation"
readme = "README.md"
authors = [
    {name = "Héctor Bastidas", email = "hector@cloudwerx.tech"},
]
dependencies = [
    "google-adk>=1.15.0,<2.0.0",
    "google-genai>=1.0.0",
    "google-auth>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "markdown2>=2.5.0",
    "weasyprint>=62.0",
]
requires-python = ">=3.12,<3.14"


[dependency-groups]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.23.8,<1.0.0",
    "nest-asyncio>=1.6.0,<2.0.0",
]

[project.optional-dependencies]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.hatch.build.targets.wheel]
packages = ["app"]
